(#106) Update send mail service
