+++ src/api/controllers/user.controller.ts
@@ -3,10 +3,10 @@ import { Request, Response } from 'express';
 import createHttpError from 'http-errors';
 import jwt from 'jsonwebtoken';
 import { HttpStatusCode } from '../../configs/statusCode.config';
+import { getVerificationEmailTemplate } from '../../helpers/mailTemplates';
 import useCatchAsync from '../../helpers/useCatchAsync';
-import getVerificationEmailTemplate from '../../helpers/verifyUserEmail';
 import { IUser, UserRoleEnum } from '../../types/user.type';
-import { sendVerificationEmail } from '../services/mail.service';
+import { sendMail } from '../services/mail.service';
 import * as UserService from '../services/user.service';
 import {
 	validateNewParentsData,
@@ -25,15 +25,13 @@ export const createTeacherAccount = useCatchAsync(async (req: Request, res: Resp
 		expiresIn: '7d'
 	});
 	const domain = req.protocol + '://' + req.get('host');
-	await sendVerificationEmail({
-		to: req.body.email,
-		subject: 'Kích hoạt tài khoản đăng nhập hệ thống quản lý giáo dục trường TH Bột Xuyên',
-		template: getVerificationEmailTemplate({
+	await sendMail(
+		getVerificationEmailTemplate({
 			redirectDomain: domain,
 			user: { ...req.body, role: UserRoleEnum.TEACHER },
 			token
 		})
-	});
+	);
 	return res.status(HttpStatusCode.CREATED).json(newUser);
 });
 
@@ -64,15 +62,13 @@ export const createParentsAccount = useCatchAsync(async (req: Request, res: Resp
 					const token = jwt.sign({ auth: recipient?.email }, process.env.ACCESS_TOKEN_SECRET!, {
 						expiresIn: '7d'
 					});
-					sendVerificationEmail({
-						to: recipient.email as string,
-						subject: 'Kích hoạt tài khoản đăng nhập hệ thống quản lý giáo dục trường TH Bột Xuyên',
-						template: getVerificationEmailTemplate({
+					sendMail(
+						getVerificationEmailTemplate({
 							redirectDomain: domain,
 							token: token,
-							user: recipient as Pick<IUser, 'displayName' | 'role'>
+							user: recipient as Pick<IUser, 'email' | 'displayName' | 'role'>
 						})
-					})
+					)
 						.then((info) => resolve(info))
 						.catch((error) => resolve(error));
 				})
@@ -85,28 +81,16 @@ export const createParentsAccount = useCatchAsync(async (req: Request, res: Resp
 		const token = jwt.sign({ auth: payload?.email }, process.env.ACCESS_TOKEN_SECRET!, {
 			expiresIn: '7d'
 		});
-		sendVerificationEmail({
-			to: payload?.email,
-			subject: 'Kích hoạt tài khoản đăng nhập hệ thống quản lý giáo dục trường TH Bột Xuyên',
-			template: getVerificationEmailTemplate({
+		sendMail(
+			getVerificationEmailTemplate({
 				redirectDomain: domain,
 				token: token,
 				user: payload
 			})
-		});
+		);
 	}
 
 	return res.status(HttpStatusCode.CREATED).json(newParents);
-});
-
-// [PATCH] /user-update
-export const updateUserInfo = useCatchAsync(async (req: Request, res: Response) => {
-	const { error, value } = validateUpdateUserData(req.body);
-	if (error) throw createHttpError.BadRequest(error.message);
-	const profile = req.profile as Pick<IUser, '_id'>;
-	const updatedUser = await UserService.updateUserInfo(profile, value);
-	if (!updatedUser) throw createHttpError.BadRequest('User does not exist!');
-	return res.status(HttpStatusCode.CREATED).json(updatedUser);
 });
 
 // [PATCH] /users/:id/update-teacher
@@ -123,9 +107,7 @@ export const updateTeacherInfo = useCatchAsync(async (req: Request, res: Respons
 export const getTeachersByStatus = useCatchAsync(async (req: Request, res: Response) => {
 	const status = req.query._status;
 	const teachers = await UserService.getTeacherUsersByStatus(status as string | undefined);
-	if (!teachers) {
-		throw createHttpError.NotFound('Không thể tìm thấy giáo viên nào!');
-	}
+	if (!teachers) throw createHttpError.NotFound('Không thể tìm thấy giáo viên nào!');
 	return res.status(HttpStatusCode.OK).json(teachers);
 });
 
@@ -141,7 +123,6 @@ export const deactivateTeacherAccount = useCatchAsync(async (req: Request, res: 
 // [GET] /users/parents/:classId
 export const getParentsUserByClass = useCatchAsync(async (req: Request, res: Response) => {
 	const parents = await UserService.getParentsUserByClass(req.params.classId);
-
 	return res.status(HttpStatusCode.OK).json(parents);
 });
 
@@ -149,7 +130,6 @@ export const getParentsUserByClass = useCatchAsync(async (req: Request, res: Res
 export const searchParentsUsers = useCatchAsync(async (req: Request, res: Response) => {
 	const result = await UserService.searchParents(req.body.searchTerm);
 	if (!result) throw createHttpError.NotFound('Cannot find any parents account!');
-
 	return res.status(HttpStatusCode.OK).json(result);
 });
 
@@ -158,6 +138,16 @@ export const getUserDetails = useCatchAsync(async (req: Request, res: Response) 
 	const user = await UserService.getUserDetails(req.params.userId);
 	if (!user) throw createHttpError.NotFound('User not found!');
 	return res.status(HttpStatusCode.OK).json(user);
+});
+
+// [PATCH] /user-update
+export const updateUserInfo = useCatchAsync(async (req: Request, res: Response) => {
+	const { error, value } = validateUpdateUserData(req.body);
+	if (error) throw createHttpError.BadRequest(error.message);
+	const profile = req.profile as Pick<IUser, '_id'>;
+	const updatedUser = await UserService.updateUserInfo(profile, value);
+	if (!updatedUser) throw createHttpError.BadRequest('User does not exist!');
+	return res.status(HttpStatusCode.CREATED).json(updatedUser);
 });
 
 // [PATCH] /users/parents/:userId
